+++ RTC_MESUREMENT_MACHINE_0_1/WrapperButton.h
@@ -12,8 +12,8 @@
 #endif
 
 class WrapperButton{
-  public: WrapperButton(int initialValue, Button *increment, Button *decrement);//un costruttore che consente di inizializzare il valore ad una certa soglia e in più che acquisisca i puntatori dei bottoni da utilizzare
-            int value;//devo esporre pubblicamente lo stato di una variabile
+  public: WrapperButton(double initialValue, Button *increment, Button *decrement);//un costruttore che consente di inizializzare il valore ad una certa soglia e in più che acquisisca i puntatori dei bottoni da utilizzare
+            double value;//devo esporre pubblicamente lo stato di una variabile
           //posso scegliere una funzione che mi va sempre a richiamare i bottoni (quindi pesa sul processore) e ritorna un valore
           //oppure posso scegliere una funzione meno prona ad errori ma che mi consente di dividere la parte di aggiornamento del valore ad una di lettura
           void checkChange();
